Новый файл для данных и мелкие правки
